Move chroniclequeue classes into specific package
