+++ src/main/scala/trw/dbsubsetter/workflow/offheap/OffHeapFkTaskQueueFactory.scala
@@ -2,7 +2,7 @@ package trw.dbsubsetter.workflow.offheap
 
 import trw.dbsubsetter.config.Config
 import trw.dbsubsetter.db.SchemaInfo
-import trw.dbsubsetter.workflow.offheap.impl.FkTaskChronicleQueue
+import trw.dbsubsetter.workflow.offheap.impl.chroniclequeue.FkTaskChronicleQueue
 
 object OffHeapFkTaskQueueFactory {
   def buildOffHeapFkTaskQueue(config: Config, schemaInfo: SchemaInfo): OffHeapFkTaskQueue = {
